repair findCoveringProgramPart()
after the latest optimizations, important (and VERY expensive) precomputations are not performed
anymore, so that findCoveringProgramPart() always returned null. Now, the program part is tried
to be found by using only local checks (on the node itself or on nodes in its neighborhood). This should not only deliver a result (at least for the cases of interest) but also be much cheaper.