wrap special case context interpreter into fallback context interpreter
this avoids crashes if a non-shrike frontend is used (e.g. the dalvik
frontend) and the additional context interpreter does not handle all non-shrike classes

This commit is a completion of 5299666039ef817f9bcc5f04f68b7fb2ec2508ff
in which the case of a non-null additional context interpreter was not
considered.