provide call graph builder factory for custom call graph builder
A call graph builder factory takes parameters such as analysis options
and scope etc.
By giving the user the opportunity to provide such a call graph builder
factory, he is able to customize the call graph builder while having
the same parameters available which the other call graph builders
also have. This way, consistency is ensured and not hoped for.