fix recognition of exit blocks when adding control flow
The node group for the exit block also contains the static fields
written by the method. Thus, the first item of this group will only
be the exit node if no static fields are written.

This leads to unsoundness in the presence of static field
writes: the exceptional out node is then not control-dependent on exception
throwing instructions anymore.

Fixing this bug raised the number of found flows slightly in one test.