a hopefully faster iRLSOD checker
There is definitely a speed-up! The complete analysis now takes about 14
minutes instead of "I don't know how many hours".

Yet, at the moment it is not really clear whether the speed-up comes
from the massive caching which was introduced alongside, or from the
improved worklist-I-do-not-more-than-I-need approach. Probably it is
both.