fix stability bug reported by simon in 7dba195
The cause of the problem was that the local killing definition analysis
used two different sources of ssa instructions:
* the array from the IR
* the instructions which occur in an (somewhat optimized) CFG
KillDef relied on the invariant that every ssa instruction in the IR also
occurs in the CFG because it first builds up field accesses from
the IR and then maps theses accesses to basic blocks in the CFG. An
optimized CFG does however not contain every instruction from the IR
which caused the analysis to crash.
The fix is to only consider instructions which are also in the CFG in
the first place.

Additionally, the killdef analysis now uses the CFG optimization
which is determined by the overall configuration.