Adapt expected numbers of violations to recent changes
Since we have in the standard case 0-1-CFA +object-sensitivity
for thread classes now (and not object-sensitivity for thread classes
and 1-CFA for the rest), there are less call graph nodes
for methods containing sources/sinks or for methods on the way.
- If there are more call graph nodes for sources/sinks, the overall
number of violations should go up: There are more sources and sinks
and hence more possible pairs. Dually, the number should go down
if we stop cloning methods which contain sources/sinks
- If there are more call graph nodes for methods which are on the
way between sources and sinks, we expect the number of violations
to tend to go down: More paths are distinguished and hence it appears
to be more likely for a given flow to disappear. Dually, if we stop
cloning such an 'in-slice method', the number of violations should
go down.