conservatively support multiple annotations of a single node
Since fields of parameters are supported, it can happen that a single
SDG node is annotated multiple times - a field of a parameter may be
annotated if the parameter is annotated and it can also be annotated
explicitly now.

If this happens, we conservatively take the most restrictive level
as annotation. For sources, we take the least upper bound of all
the multiple annotations, for sinks we take the greatest lower bounds.

Note that this is kind of a quick fix. The root cause of this strange
behaviour is that there is no possibility to annotate exactly
a non-primitive root parameter node using JOANA's API automagic.