WALA: make element types in array accesses more precise
Because it cannot be seen directly from the Java bytecode
instruction, element types for array accesses are assumed
to be java.lang.Object if they are not primitive
(there is a distinct array access instruction
for each primitive type, those access do not cause problems).

This is improved now by using type inference information to
refine the element types.

As the aliasing analysis uses (among other stuff) equality checks
on types to determine aliasing and more array element types are
equal now, this led to a bit more aliasing and hence a tiny bit
more violations in one test case. The expected number of violations
has been increased to remedy this.