fix in WALA: JarStreamModule must arbitrary InputStreams
This change is neccessary, since JarStreamModule inherits from
JarInputStream now and the constructor of JarInputStream takes the
stream to extract and not the stream to take data directly from.

The usages of JarStreamModule also have been adapted: We can get rid of
the intermediate JarInputStream now, since the JarStreamModule we pass
the stream to is also a JarInputStream (and therefore extracts data,
which was done by the intermediate layer before).